refactor(i18n): reorg translations by feature, not locale
