+++ frontend/src/i18n/locales/en-sg.ts
@@ -1,6 +1,5 @@
-import Translation from '../types'
-
-import { login } from './features'
+import { enSG as login } from './features/login'
+import Translation from './types'
 
 export const enSG: Translation = {
   translation: {
