--- frontend/src/i18n/locales/en-sg/features/index.ts
@@ -1,1 +1,0 @@
-export * from './login'
