+++ frontend/src/i18n/locales/features/login/index.ts
@@ -1,0 +1,16 @@
+export * from './en-sg'
+export * from './zh-sg'
+
+export interface Login {
+  components: {
+    LoginForm: {
+      onlyAvailableForPublicOfficers: string
+      emailEmptyErrorMsg: string
+      login: string
+      haveAQuestion: string
+    }
+  }
+  LoginPage: {
+    slogan: string
+  }
+}
