+++ frontend/src/i18n/locales/types.ts
@@ -1,19 +1,9 @@
+import { Login } from './features/login'
+
 interface Translation {
   translation: {
     features: {
-      login: {
-        components: {
-          LoginForm: {
-            onlyAvailableForPublicOfficers: string
-            emailEmptyErrorMsg: string
-            login: string
-            haveAQuestion: string
-          }
-        }
-        LoginPage: {
-          slogan: string
-        }
-      }
+      login: Login
     }
   }
 }
