--- frontend/src/i18n/locales/zh-sg/features/index.ts
@@ -1,1 +1,0 @@
-export * from './login'
